requirements.txt
Flask==3.0.0
opencv-python==4.9.0.80
numpy==1.26.2
Pillow==10.1.0
mediapipe==0.10.21
Werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
